Demonstration/requirements.txt
Flask==1.1.2
matplotlib==3.2.2
scipy==1.7.0
pandas==1.3.0
numpy==1.19.5
tensorflow==2.5.0
joblib==1.0.1
h5py==3.1.0
plotly==5.1.0
PyYAML==5.4.1
scikit_learn==0.22.2.post1
